py3_requirements.txt
## How to set up a Python runtime environment for the Whole Cell Model.
## (See docs/create-pyenv.md for more details, esp. for Sherlock.)
## (See docs/dev-tools.md for prerequisites such as installing command line
## tools, additional SDK headers, and add pyenv init to your shell.)
#
## Install these packages using your local package manager such as homebrew on macOS:
#   brew install pyenv pyenv-virtualenv pyenv-virtualenvwrapper
#   brew install glpk openssl readline swig suite-sparse xz
#
## Whole Cell Model requires OpenBLAS 0.3.7+ or equivalent. Installing numpy and
## scipy binary wheels now provides a suitable OpenBLAS by default.
## (OpenBLAS before 0.3.7 had threading bugs. So does Apple's "Accelerate"
## framework. We haven't tested recent releases of Intel Math Kernel Library.)
#
## Install the required version of Python via pyenv, and remember to enable it as
## a shared library:
#   PYTHON_CONFIGURE_OPTS="--enable-shared" pyenv install 3.8.3
#
## Create the "wcEcoli3" python virtual environment based on that Python version,
## and select it for your project directory:
#   cd ~/dev/wcEcoli
#   pyenv virtualenv 3.8.3 wcEcoli3 && pyenv local wcEcoli3
#
## Upgrade this virtual environment's installers:
#   pip install --upgrade pip setuptools virtualenv virtualenvwrapper virtualenv-clone wheel
#
## Install numpy (if it's the first time) then all the packages listed in this file:
#   pip install numpy==1.19.0
#   pip install -r py3_requirements.txt && pyenv rehash
#
## Configure the installed matplotlib per the instructions in docs/create-pyenv.md

# Installers
pip>=20.1.1
setuptools>=49.1.0
virtualenv>=20.0.26
virtualenv-clone>=0.5.4
virtualenvwrapper>=4.8.4
wheel>=0.34.2

# Direct dependendencies, frozen at the latest version 2020-07-08, except bokeh is
# frozen at the latest version that's compatible with Python 2.
#
# TODO: Periodically unfreeze, update, test, and refreeze.
numpy==1.19.0
scipy==1.5.1

biopython==1.77
bokeh==1.4.0
borealis-fireworks==0.6.6
confluent-kafka==1.4.2
cvxpy==1.1.1
Cython==0.29.21
Equation==1.2.1
FireWorks==1.9.5
future==0.18.2
ipdb==0.13.3
ipython==7.16.1
line-profiler==3.0.2
llvmlite==0.33.0
matplotlib==3.2.2
mock==4.0.2
numba==0.50.1
pandas==1.0.5
Pillow==7.2.0
pymongo==3.10.1
pytest==5.4.3
PyYAML==5.3.1
ruamel.yaml==0.16.10
six==1.15.0
stochastic-arrow==0.4.1
swiglpk==4.65.1
sympy==1.6.1
Theano==1.0.4
Unum==4.1.4
wholecell-vivarium==0.0.52

## The following requirements were added by pip freeze:
appdirs==1.4.4
appnope==0.1.0
Arpeggio==1.9.2
attrs==19.3.0
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cobra==0.18.1
cycler==0.10.0
decorator==4.4.2
depinfo==1.5.3
distlib==0.3.1
docker==4.2.2
ecos==2.0.7.post1
filelock==3.0.12
Flask==1.1.2
flask-paginate==0.7.0
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-cloud-logging==1.15.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
gunicorn==20.0.4
idna==2.10
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
monty==3.0.2
more-itertools==8.4.0
mpmath==1.1.0
networkx==2.4
optlang==1.4.4
osqp==0.6.1
packaging==20.4
parsimonious==0.8.1
parso==0.7.0
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pint==0.14
pipdeptree==1.0.0
pluggy==0.13.1
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
py==1.9.0
py-cpuinfo==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pymunk==5.6.0
pyparsing==2.4.7
pytest-benchmark==3.2.3
python-dateutil==2.8.1
python-libsbml-experimental==5.18.1
pytz==2020.1
requests==2.24.0
rsa==4.6
ruamel.yaml.clib==0.2.0
scs==2.1.2
stevedore==2.0.1
subprocess32==3.5.4
tabulate==0.8.7
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
